tier0-tier1: Fix compilation errors
Tested on clang with `make tier1`

+++ public/tier1/characterset.h
@@ -37,7 +37,7 @@ extern void CharacterSetBuild( characterset_t *pSetBuffer, const char *pSetStrin
 //			character - character to lookup
 // Output : int - 1 if the character was in the set
 //-----------------------------------------------------------------------------
-#define IN_CHARACTERSET( SetBuffer, character )		((SetBuffer).set[(character)])
+#define IN_CHARACTERSET( SetBuffer, character )		((SetBuffer).set[(int)(character)])
 
 
 #endif // CHARACTERSET_H
